backend/requirements.txt
requests
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
python-jose
passlib
python-multipart
